make components and reorganise branch
