+++ frontend/components/CourseEvaluation/SectionTabs.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { useSWRAuth } from '@/components/hooks/useSWRAuth';
 import type { NextPage } from 'next';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
@@ -12,17 +11,8 @@ import Typography from '@mui/material/Typography';
 import styles from '../styles/Home.module.css';
 import { useContext } from 'react';
 import { useRouter } from 'next/router';
-import AppContext from 'components/Context/TopLevelContext';
-import useAuthenticatedAPIClient from '@/components/hooks/useAuthenticatedAPIClient';
-import { determineIfUserIsAuthentication } from 'utils/Authentication';
-import Overview from '@/components/course-evaluation/overview';
-import {
-  API_ENDPOINT,
-  CourseEvaluationListEntry,
-  DEFAULT_COURSE_EVALUATION_LIST_ENTRY,
-  DEFAULT_USER_API_RESPONSE,
-  UserAPIResponse,
-} from 'utils/api';
+
+import Overview from './Overview';
 
 interface TabPanelProps {
   children?: React.ReactNode;
@@ -57,21 +47,13 @@ function a11yProps(index: number) {
   };
 }
 
-const Coordinator_evaluation: NextPage = () => {
+const SectionTabs: NextPage = () => {
   // Change these data when this issue gets worked on. For now it just contains a demo of how to get user info
-  const { response, isLoading, error } = useSWRAuth(API_ENDPOINT.COURSE_EVALUATION.LIST);
-  const courseEvaluationListEntries = ((response?.data as unknown) ||
-    []) as CourseEvaluationListEntry[];
 
-  const router = useRouter();
-  const axios = useAuthenticatedAPIClient();
-  const { authenticationDetails, setAuthenticationDetails } = useContext(AppContext);
-  const isUserAuthenticated = determineIfUserIsAuthentication(authenticationDetails.accessToken);
-
-  const [value, setValue] = React.useState(0);
+  const [tabsValue, setTabsValue] = React.useState(0);
 
   const handleChangeTab = (event: React.SyntheticEvent, newValue: number) => {
-    setValue(newValue);
+    setTabsValue(newValue);
   };
 
   return (
@@ -86,7 +68,7 @@ const Coordinator_evaluation: NextPage = () => {
       <Box
         sx={{ borderBottom: 1, borderColor: 'divider', backgroundColor: '#0E91AC', width: '90%' }}
       >
-        <Tabs value={value} onChange={handleChangeTab} variant="fullWidth">
+        <Tabs value={tabsValue} onChange={handleChangeTab} variant="fullWidth">
           <Tab label="OVERVIEW" {...a11yProps(0)} />
           <Tab label="JUSTIFICATIONS" {...a11yProps(1)} />
           <Tab label="DOCUMENTS" {...a11yProps(2)} />
@@ -94,16 +76,16 @@ const Coordinator_evaluation: NextPage = () => {
         </Tabs>
       </Box>
       <Box sx={{ backgroundColor: '#EEEEEE', width: '100%' }}>
-        <TabPanel value={value} index={0}>
+        <TabPanel value={tabsValue} index={0}>
           <Overview />
         </TabPanel>
-        <TabPanel value={value} index={1}>
+        <TabPanel value={tabsValue} index={1}>
           Item Two
         </TabPanel>
-        <TabPanel value={value} index={2}>
+        <TabPanel value={tabsValue} index={2}>
           Item Three
         </TabPanel>
-        <TabPanel value={value} index={3}>
+        <TabPanel value={tabsValue} index={3}>
           Item Four
         </TabPanel>
       </Box>
@@ -111,4 +93,4 @@ const Coordinator_evaluation: NextPage = () => {
   );
 };
 
-export default Coordinator_evaluation;
+export default SectionTabs;
