+++ frontend/pages/course-evaluation/[id].js
@@ -2,6 +2,7 @@ import { useRouter } from 'next/router';
 import Card from '@mui/material/Card';
 import Container from '@mui/material/Container';
 import Typography from '@mui/material/Typography';
+import SectionTabs from '@/components/CourseEvaluation/SectionTabs'
 
 
 const Post = () => {
@@ -9,13 +10,9 @@ const Post = () => {
   const { id } = router.query;
 
   console.log(id)
-
+  
   return (
-    <Container>
-      <Card>
-        <Typography>hello</Typography>
-      </Card>
-    </Container>
+    <SectionTabs />
   );
 };
